Add layout and index for redirecting /amqp.node/
GitHub pages will redirect <name>.github.io/<repo> to a CNAME, if you
also have a <name>.github.io repo. This means they _won't_ get
redirected to the new location (after moving the repo)
https://amqp-node.github.io/amqplib/.

This commit adds a layout which will serve a meta-refresh, and a test
that it works for /amqp.node/. This is less good than sending a proper
HTTP 301, and has to be done for each page, but will do until I can
move off GitHub pages for www.squaremobius.net.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>